# backend/contracts/openapi.mvp.yaml
openapi: 3.0.3
info:
  title: AgriNextGen MVP API Contracts
  version: 0.1.0
  description: Minimal contract-first API for MVP implementation.
servers:
  - url: http://localhost:4000
...
paths:
...
  /auth/send-otp:
    post:
      tags: [auth]
      summary: Send OTP to user phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSentResponse"
  /auth/verify-otp:
    post:
      tags: [auth]
      summary: Verify OTP and return tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Auth successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
components:
...
  schemas:
    SendOtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    VerifyOtpRequest:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    OtpSentResponse:
      type: object
      properties:
        status:
          type: string
        expires_in_seconds:
          type: integer
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in_seconds:
          type: integer
        user:
          $ref: "#/components/schemas/Profile"
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        role:
          type: string
        phone:
          type: string
...